Put `approx_bytes` into `table_slice` and cache it
