+++ libtenzir/include/tenzir/arrow_table_slice.hpp
@@ -42,6 +42,9 @@ struct arrow_table_slice_state<fbs::table_slice::arrow::v2> {
 
   auto get_flat_columns() const -> const arrow::ArrayVector&;
 
+  mutable std::atomic<uint64_t> approx_bytes_
+    = std::numeric_limits<uint64_t>::max();
+
   /// Whether the record batch points to outside data.
   bool is_serialized;
 };
@@ -118,6 +121,8 @@ public:
   /// @returns A shared pointer to the underlying Arrow Record Batch.
   [[nodiscard]] std::shared_ptr<arrow::RecordBatch>
   record_batch() const noexcept;
+
+  auto approx_bytes() const -> uint64_t;
 
 private:
   // -- implementation details -------------------------------------------------
